# openapi.v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Relay Knowledge API",
    "description": "File ingestion and vector search with JWT+RLS+AAD security",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v2/knowledge/upload": {
      "post": {
        "tags": [
          "Knowledge API"
        ],
        "summary": "Upload File",
        "description": "Upload a file for knowledge ingestion.\nReturns 202 Accepted with file_id; processing happens asynchronously.\n\nSecurity:\n- JWT validation required\n- File size limit: 50MB\n- MIME type whitelist enforced\n- Metadata AAD-encrypted with user_hash binding\n- RLS enforced at insert time",
        "operationId": "upload_file_api_v2_knowledge_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v2_knowledge_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/knowledge/index": {
      "post": {
        "tags": [
          "Knowledge API"
        ],
        "summary": "Index File",
        "description": "Index a file: extract \u2192 chunk \u2192 embed \u2192 store vectors.\n\nSecurity:\n- JWT + RLS (owner-only check)\n- AAD binding on metadata\n- Circuit breaker for embedding service failure",
        "operationId": "index_file_api_v2_knowledge_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileIndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileIndexResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/knowledge/search": {
      "post": {
        "tags": [
          "Knowledge API"
        ],
        "summary": "Search Knowledge",
        "description": "Vector similarity search over indexed embeddings.\n\nSecurity:\n- JWT + RLS (only own files visible)\n- AAD verification on metadata decryption",
        "operationId": "search_knowledge_api_v2_knowledge_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/knowledge/files": {
      "get": {
        "tags": [
          "Knowledge API"
        ],
        "summary": "List Files",
        "description": "List user's files (RLS-isolated).\n\nSecurity:\n- JWT + RLS (only own files returned)",
        "operationId": "list_files_api_v2_knowledge_files_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/knowledge/files/{file_id}": {
      "delete": {
        "tags": [
          "Knowledge API"
        ],
        "summary": "Delete File",
        "description": "Delete file and cascade delete embeddings.\n\nSecurity:\n- JWT + RLS (owner-only)\n- Explicit ownership check before cascade delete\n- AAD verification before deletion",
        "operationId": "delete_file_api_v2_knowledge_files__file_id__delete",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Knowledge API"
        ],
        "summary": "Handle Invalid Endpoint",
        "description": "Catch unsupported operations",
        "operationId": "handle_invalid_endpoint_api_v2_knowledge_files__file_id__post",
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_file_api_v2_knowledge_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "default": "upload"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "default": "[]"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_v2_knowledge_upload_post"
      },
...
      "FileIndexRequest": {
        "properties": {
          "file_id": {
            "type": "string",
            "format": "uuid",
            "title": "File Id",
            "description": "File ID from upload response"
          },
          "chunk_strategy": {
            "$ref": "#/components/schemas/ChunkStrategy",
            "description": "Text chunking strategy",
            "default": "smart"
          },
          "chunk_overlap": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Chunk Overlap",
            "description": "Overlap between chunks in tokens",
            "default": 100
          },
          "embedding_model": {
            "$ref": "#/components/schemas/EmbeddingModel",
            "description": "Embedding model to use",
            "default": "ada-002"
          },
          "rerank": {
            "type": "boolean",
            "title": "Rerank",
            "description": "Enable reranking",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "file_id"
        ],
        "title": "FileIndexRequest",
        "description": "Request model for POST /api/v2/knowledge/index",
        "example": {
          "chunk_overlap": 100,
          "chunk_strategy": "smart",
          "embedding_model": "ada-002",
          "file_id": "550e8400-e29b-41d4-a716-446655440000",
          "rerank": true
        }
      },
      "FileIndexResponse": {
        "properties": {
          "file_id": {
            "type": "string",
            "format": "uuid",
            "title": "File Id",
            "description": "File ID indexed"
          },
          "chunks_created": {
            "type": "integer",
            "title": "Chunks Created",
            "description": "Number of chunks created"
          },
          "tokens_processed": {
            "type": "integer",
            "title": "Tokens Processed",
            "description": "Total tokens processed"
          },
          "embedding_latency_ms": {
            "type": "integer",
            "title": "Embedding Latency Ms",
            "description": "Embedding generation latency"
          },
          "embedding_model_used": {
            "type": "string",
            "title": "Embedding Model Used",
            "description": "Embedding model used"
          },
          "vectors_stored": {
            "type": "integer",
            "title": "Vectors Stored",
            "description": "Vectors successfully stored"
          },
          "file_url": {
            "type": "string",
            "title": "File Url",
            "description": "URL to access file metadata"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Indexing status",
            "default": "indexed"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "chunks_created",
          "tokens_processed",
          "embedding_latency_ms",
          "embedding_model_used",
          "vectors_stored",
          "file_url"
        ],
        "title": "FileIndexResponse",
        "description": "Response for POST /api/v2/knowledge/index (200 OK)"
      },
...
      "FileUploadResponse": {
        "properties": {
          "file_id": {
            "type": "string",
            "format": "uuid",
            "title": "File Id",
            "description": "Unique file ID"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Processing status",
            "default": "queued"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id",
            "description": "Request ID for tracing"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Status message",
            "default": "File queued for processing"
          },
          "expected_completion_ms": {
            "type": "integer",
            "title": "Expected Completion Ms",
            "description": "Estimated completion time (ms)",
            "default": 5000
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "request_id"
        ],
        "title": "FileUploadResponse",
        "description": "Response for POST /api/v2/knowledge/upload (202 Accepted)"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Original query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "type": "array",
            "title": "Results",
            "description": "Search results"
          },
          "total_results": {
            "type": "integer",
            "title": "Total Results",
            "description": "Total matching results"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms",
            "description": "Query latency (ms)"
          },
          "embedding_model_used": {
            "type": "string",
            "title": "Embedding Model Used",
            "description": "Embedding model used"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "description": "Whether result was from cache",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "query",
          "results",
          "total_results",
          "latency_ms",
          "embedding_model_used"
        ],
        "title": "SearchResponse",
        "description": "Response for POST /api/v2/knowledge/search (200 OK)"
      },
...
    }
  }
}